fix: Permission issues

+++ Plugin/src/main/java/de/corneliusmay/silkspawners/plugin/listeners/BlockBreakListener.java
@@ -24,7 +24,7 @@ public class BlockBreakListener extends SilkSpawnersListener<BlockBreakEvent> {
         if(!spawner.isValid()) return;
 
         Player p = e.getPlayer();
-        if(!p.hasPermission("silkspawners.break." + spawner.getEntityType().getName())) {
+        if(!p.hasPermission("silkspawners.break." + spawner.getEntityType().getName()) && !p.hasPermission("silkspawners.break.*")) {
             destroySpawner(p, e);
             return;
         }
